+++ example/android/app/src/main/kotlin/np/com/sarbagyastha/nepali_date_picker_example/MainActivity.kt
@@ -2,4 +2,4 @@ package np.com.sarbagyastha.nepali_date_picker_example
 
 import io.flutter.embedding.android.FlutterActivity
 
-class MainActivity: FlutterActivity()
+class MainActivity : FlutterActivity()