+++ example/lib/app_scope.dart
@@ -2,22 +2,18 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import 'package:flutter/material.dart';
+import 'package:material_ui/material_ui.dart';
 import 'package:nepali_utils/nepali_utils.dart';
 
 ///
-class _AppModel extends ChangeNotifier {
-  ///
-  _AppModel(Locale locale, Brightness brightness, Color color)
-    : _color = color,
-      _locale = locale,
-      _brightness = brightness {
+class _AppModel(
+  var Locale _locale,
+  var Brightness _brightness,
+  var Color _color,
+) extends ChangeNotifier {
+  this {
     _updateLanguage();
   }
-
-  Locale _locale;
-  Brightness _brightness;
-  Color _color;
 
   ///
   void toggleLocale() {
@@ -52,35 +48,26 @@ class _AppModel extends ChangeNotifier {
   Color get color => _color;
 
   void _updateLanguage() {
-    NepaliUtils().language = _locale.languageCode == 'ne'
-        ? Language.nepali
-        : Language.english;
+    NepaliUtils().language = _locale.languageCode == 'ne' ? .nepali : .english;
   }
 }
 
 ///
-class AppScope extends StatefulWidget {
+class const AppScope({
   ///
-  const AppScope({
-    required this.builder,
-    this.defaultLocale = const Locale('en', 'US'),
-    this.defaultBrightness = Brightness.light,
-    this.defaultColor = Colors.orange,
-    super.key,
-  });
+  required final Widget Function(BuildContext, Locale, Brightness, Color)
+  builder,
 
   ///
-  final Widget Function(BuildContext, Locale, Brightness, Color) builder;
+  final Locale defaultLocale = const Locale('en', 'US'),
 
   ///
-  final Locale defaultLocale;
+  final Brightness defaultBrightness = .light,
 
   ///
-  final Brightness defaultBrightness;
-
-  ///
-  final Color defaultColor;
-
+  final Color defaultColor = Colors.teal,
+  super.key,
+}) extends StatefulWidget {
   ///
   // ignore: library_private_types_in_public_api
   static _AppModel of(BuildContext context) {
@@ -131,11 +118,8 @@ class _AppScopeState extends State<AppScope> {
   }
 }
 
-class _LocaleScope extends InheritedWidget {
-  const _LocaleScope({required this.model, required super.child});
-
-  final _AppModel model;
-
+class const _LocaleScope({required final _AppModel model, required super.child})
+    extends InheritedWidget {
   @override
   bool updateShouldNotify(_LocaleScope old) {
     return model._locale != old.model._locale ||