+++ example/lib/modes/calendar_date_range_picker_widget.dart
@@ -2,13 +2,11 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import 'package:flutter/material.dart';
+import 'package:material_ui/material_ui.dart';
 import 'package:nepali_date_picker/nepali_date_picker.dart';
 
 ///
-class CalendarDateRangePickerWidget extends StatefulWidget {
-  const CalendarDateRangePickerWidget({super.key});
-
+class const CalendarDateRangePickerWidget({super.key}) extends StatefulWidget {
   @override
   State<CalendarDateRangePickerWidget> createState() =>
       _CalendarDateRangePickerWidgetState();
@@ -24,37 +22,35 @@ class _CalendarDateRangePickerWidgetState
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      body: Column(
-        children: [
-          Expanded(
-            child: CalendarDateRangePicker(
-              initialStartDate: NepaliDateTime.now(),
-              initialEndDate: NepaliDateTime.now().add(Duration(days: 5)),
-              firstDate: NepaliDateTime(1970),
-              lastDate: NepaliDateTime(2100),
-              onStartDateChanged: (date) {
-                _dateRange = (date as NepaliDateTime, _dateRange.$2);
-                setState(() {});
-              },
-              onEndDateChanged: (date) {
-                _dateRange = (_dateRange.$1, date as NepaliDateTime?);
-                setState(() {});
-              },
-              selectableDayPredicate: null,
-              calendarDelegate: const NepaliCalendarDelegate(),
-            ),
+    return Column(
+      children: [
+        Expanded(
+          child: CalendarDateRangePicker(
+            initialStartDate: NepaliDateTime.now(),
+            initialEndDate: NepaliDateTime.now().add(Duration(days: 5)),
+            firstDate: NepaliDateTime(1970),
+            lastDate: NepaliDateTime(2100),
+            onStartDateChanged: (date) {
+              _dateRange = (date as NepaliDateTime, _dateRange.$2);
+              setState(() {});
+            },
+            onEndDateChanged: (date) {
+              _dateRange = (_dateRange.$1, date as NepaliDateTime?);
+              setState(() {});
+            },
+            selectableDayPredicate: null,
+            calendarDelegate: const NepaliCalendarDelegate(),
           ),
-          ListTile(
-            shape: RoundedRectangleBorder(
-              borderRadius: BorderRadius.vertical(bottom: Radius.circular(16)),
-            ),
-            title: Text('From: ${_format(_dateRange.$1)}'),
-            subtitle: Text('To: ${_format(_dateRange.$2)}'),
-            tileColor: Theme.of(context).colorScheme.primaryContainer,
+        ),
+        ListTile(
+          shape: RoundedRectangleBorder(
+            borderRadius: BorderRadius.vertical(bottom: Radius.circular(16)),
           ),
-        ],
-      ),
+          title: Text('From: ${_format(_dateRange.$1)}'),
+          subtitle: Text('To: ${_format(_dateRange.$2)}'),
+          tileColor: Theme.of(context).colorScheme.primaryContainer,
+        ),
+      ],
     );
   }