+++ example/lib/modes/date_range_picker_widget.dart
@@ -2,14 +2,12 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import 'package:flutter/material.dart';
+import 'package:material_ui/material_ui.dart';
 import 'package:intl/intl.dart';
 import 'package:nepali_date_picker/nepali_date_picker.dart';
 
 ///
-class DateRangePickerWidget extends StatefulWidget {
-  const DateRangePickerWidget({super.key});
-
+class const DateRangePickerWidget({super.key}) extends StatefulWidget {
   @override
   State<DateRangePickerWidget> createState() => _DateRangePickerWidgetState();
 }
@@ -22,72 +20,64 @@ class _DateRangePickerWidgetState extends State<DateRangePickerWidget> {
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      body: Padding(
-        padding: const EdgeInsets.all(16),
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          crossAxisAlignment: CrossAxisAlignment.stretch,
-          children: [
-            Card(
-              margin: EdgeInsets.zero,
-              child: Padding(
-                padding: const EdgeInsets.symmetric(
-                  vertical: 16,
-                  horizontal: 8,
-                ),
-                child: _selectedDateTimeRange == null
-                    ? Text('No Date Range Picked!', textAlign: TextAlign.center)
-                    : Column(
-                        spacing: 16,
-                        children: [
-                          Text(
-                            NepaliDateFormat(
-                              'EEE, MMMM d, y',
-                            ).format(_selectedDateTimeRange!.start),
-                            style: Theme.of(context).textTheme.titleLarge,
-                            textAlign: TextAlign.center,
-                          ),
-                          Text(
-                            DateFormat('EEE, MMMM d, y').format(
-                              _selectedDateTimeRange!.start.toDateTime(),
-                            ),
-                            style: Theme.of(context).textTheme.titleSmall,
-                            textAlign: TextAlign.center,
-                          ),
-                          Icon(Icons.arrow_downward_rounded),
-                          Text(
-                            NepaliDateFormat(
-                              'EEE, MMMM d, y',
-                            ).format(_selectedDateTimeRange!.end),
-                            style: Theme.of(context).textTheme.titleLarge,
-                            textAlign: TextAlign.center,
-                          ),
-                          Text(
-                            DateFormat(
-                              'EEE, MMMM d, y',
-                            ).format(_selectedDateTimeRange!.end.toDateTime()),
-                            style: Theme.of(context).textTheme.titleSmall,
-                            textAlign: TextAlign.center,
-                          ),
-                        ],
-                      ),
-              ),
+    return Padding(
+      padding: const EdgeInsets.all(16),
+      child: Column(
+        mainAxisAlignment: .center,
+        crossAxisAlignment: .stretch,
+        children: [
+          Card(
+            margin: EdgeInsets.zero,
+            child: Padding(
+              padding: const EdgeInsets.symmetric(vertical: 16, horizontal: 8),
+              child: _selectedDateTimeRange == null
+                  ? Text('No Date Range Picked!', textAlign: .center)
+                  : Column(
+                      spacing: 16,
+                      children: [
+                        Text(
+                          NepaliDateFormat('EEE, MMMM d, y')
+                              .format(_selectedDateTimeRange!.start),
+                          style: Theme.of(context).textTheme.titleLarge,
+                          textAlign: .center,
+                        ),
+                        Text(
+                          DateFormat(
+                            'EEE, MMMM d, y',
+                          ).format(_selectedDateTimeRange!.start.toDateTime()),
+                          style: Theme.of(context).textTheme.titleSmall,
+                          textAlign: .center,
+                        ),
+                        Icon(Icons.arrow_downward_rounded),
+                        Text(
+                          NepaliDateFormat('EEE, MMMM d, y')
+                              .format(_selectedDateTimeRange!.end),
+                          style: Theme.of(context).textTheme.titleLarge,
+                          textAlign: .center,
+                        ),
+                        Text(
+                          DateFormat('EEE, MMMM d, y')
+                              .format(_selectedDateTimeRange!.end.toDateTime()),
+                          style: Theme.of(context).textTheme.titleSmall,
+                          textAlign: .center,
+                        ),
+                      ],
+                    ),
             ),
-            SizedBox(height: 20),
-            FilledButton.tonal(
-              onPressed: () async {
-                _selectedDateTimeRange = await showNepaliDateRangePicker(
-                  context: context,
-                  firstDate: NepaliDateTime(2020),
-                  lastDate: NepaliDateTime(2099),
-                );
-                setState(() {});
-              },
-              child: Text('PICK DATE RANGE'),
-            ),
-          ],
-        ),
+          ),
+          SizedBox(height: 20),
+          FilledButton.tonal(
+            onPressed: () async {
+              _selectedDateTimeRange = await showNepaliDateRangePicker(
+                context: context,
+                firstDate: NepaliDateTime(2020),
+                lastDate: NepaliDateTime(2099),
+              );
+              setState(() {});
+            },
+            child: Text('PICK DATE RANGE'),
+          ),
+        ],
       ),
     );
   }